Now compiles with Intel 2024

+++ src/distributed_io.c
@@ -962,8 +962,10 @@ void write_output_distributed(struct engine* e,
 
   /* Compute offset in the file and total number of particles */
   long long N[swift_type_count] = {
-      Ngas_written,   Ndm_written,         Ndm_background, Nsinks_written,
-      Nstars_written, Nblackholes_written, Ndm_neutrino};
+      (long long)Ngas_written,   (long long)Ndm_written,
+      (long long)Ndm_background, (long long)Nsinks_written,
+      (long long)Nstars_written, (long long)Nblackholes_written,
+      (long long)Ndm_neutrino};
 
   /* Gather the total number of particles to write */
   long long N_total[swift_type_count] = {0};
